Implementaddo mensaje de confirmacion de eliminacion de metrica en la interfaz de usuario

+++ src/MetricaIndex/presenterVerMetrica.js
@@ -20,17 +20,21 @@ document.addEventListener("DOMContentLoaded", () => {
     const botonesEliminarMetrica = document.querySelectorAll(".eliminar-metrica");
     botonesEliminarMetrica.forEach(boton => {
     boton.addEventListener("click", function() {
-        const metricaIndex = parseInt(this.dataset.metricaIndex);
-        const metricaEliminada = proyecto.metricas[metricaIndex];
-        const eliminacionExitosa = metrica.eliminarMetricaDeProyecto(metricaEliminada, proyecto);
-        
-        if (eliminacionExitosa) {
-            this.parentNode.remove();
-            localStorage.setItem("proyectoActual", JSON.stringify(proyecto));
-            metrica.actualizarProyectoEnArray(proyecto);
-        } else {
-            console.error("Error al eliminar la métrica del proyecto");
-        }
+        if(metrica.confirmarEliminarMetrica())
+            {
+                const metricaIndex = parseInt(this.dataset.metricaIndex);
+                const metricaEliminada = proyecto.metricas[metricaIndex];
+                const eliminacionExitosa = metrica.eliminarMetricaDeProyecto(metricaEliminada, proyecto);
+                
+                if (eliminacionExitosa) {
+                    this.parentNode.remove();
+                    localStorage.setItem("proyectoActual", JSON.stringify(proyecto));
+                    metrica.actualizarProyectoEnArray(proyecto);
+                } else {
+                    console.error("Error al eliminar la métrica del proyecto");
+                }
+            }
+
     });
     });
